% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ctd}
\alias{ctd}
\title{Human Embryonic Midbrain cell type data file}
\format{An object of class \code{list} of length 2.}
\source{
The code to prepare the .Rda file file from the marker file is:
\code{
library(EWCE)
data(raw_sc_expression_data)
raw_sc_expression_data$exp_CORRECTED = fix.bad.hgnc.symbols(raw_sc_expression_data$exp)
raw_sc_expression_data$exp_CORRECTED_DROPPED = drop.uninformative.genes(exp=raw_sc_expression_data$exp_CORRECTED, level2annot = raw_sc_expression_data$annot$level2class)
fNames = generate.celltype.data(exp=raw_sc_expression_data$exp_CORRECTED_DROPPED,raw_sc_expression_data$annot,"HumanEmbyronicMidbrain")
load(fNames[1])
usethis::use_data(ctd,overwrite=TRUE)
}
}
\usage{
ctd
}
\description{
The processed data for the human embryonic midbrain, ready for use in MAGMA_Celltyping and EWCE
}
\keyword{datasets}
