% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_raw_sc_data.r
\name{process_raw_sc_data}
\alias{process_raw_sc_data}
\title{process_raw_sc_data}
\usage{
process_raw_sc_data()
}
\description{
\code{process_raw_sc_data} downloads the raw data from GEO (accession GSE76381) and formats it.
}
\examples{
raw_sc_expression_data = process_raw_sc_data()
usethis::use_data(raw_sc_expression_data,overwrite=TRUE)
}
