% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_sc_expression_data}
\alias{raw_sc_expression_data}
\title{Human Embryonic Midbrain raw expression data}
\format{An object of class \code{list} of length 3.}
\source{
The code to prepare the .Rda file file from the marker file is:
\code{
raw_sc_expression_data = process_raw_sc_data()
usethis::use_data(raw_sc_expression_data,overwrite=TRUE)
}
}
\usage{
raw_sc_expression_data
}
\description{
The raw expression and annotation data for the human embryonic midbrain. If you want to regenerate the just download the data from GSE76381. The annotation labels are customised below.
}
\keyword{datasets}
